Refactor slave code
Move classes from slave.js monolit into their own source files.
